updated dependencies

+++ app/src/main/java/com/example/movies/model/Movie.kt
@@ -6,7 +6,7 @@ data class Movie(
     val id: String,
     val title: String,
     val overview: String,
-    val voteAverage: Float,
+    val voteAverage: Double,
     val poster: String,
     val releaseDate: LocalDate,
 )
